Graded_portolio index file
Html coding for the jewelry Mangata and Gallo concept store. It contains metatags to assure it works well with the SEO and social media cards when its shown.
It contains all the basic structures within the main html tag, head, body and footer, a link to bootstrap too. 
The structure is divided by a heading with a logo and a navbar, a background image attached to the css file and the structure of the container with separated rows with some items into colums and the footer with a second logo and a text.